More refactor + stylesheet cleanup
