+++ packages/esm-patient-test-results-app/src/filter/filter-set.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import styles from './filter-set.scss';
+import styles from './filter-set.styles.scss';
 import { Accordion, AccordionItem, Checkbox } from 'carbon-components-react';
 import FilterContext from './filter-context';
 import { useConfig, useLayoutType } from '@openmrs/esm-framework';
