+++ packages/esm-patient-test-results-app/src/grouped-timeline/index.ts
@@ -1,5 +1,6 @@
 import GroupedTimeline from './grouped-timeline';
 import styles from './grouped-timeline.styles.scss';
+import useGetManyObstreeData from './useObstreeData';
 
 export default GroupedTimeline;
-export { GroupedTimeline, styles };
+export { GroupedTimeline, styles, useGetManyObstreeData };
